added not_if for execute resource in install_zookeeper

+++ recipes/install_zookeeper.rb
@@ -43,12 +43,16 @@ execute 'unzip and untar zookeeper source' do
   group 'zookeeper'
   umask 022
   action :run
+  not_if do
+    { ::Dir.exists?('/tmp/zookeeper-3.4.12') }
+  end
 end
 
 # move to /opt and rename
 execute 'move to /opt and rename' do
   command 'cp -rp /tmp/zookeeper-3.4.12 /opt/ && mv /opt/zookeeper-3.4.12 /opt/zookeeper'
   action :run
+  not_if { ::Dir.exists('/opt/zookeeper') }
 end
 
 # create required directories and files
